requirements.txt
Django==2.1
django-ckeditor==5.7.1
django-js-asset==1.2.2
pytz==2019.1
whitenoise==4.1.2
